Format code

+++ face_detection/yolov5_face/utils/autoanchor.py
@@ -5,7 +5,6 @@ import torch
 import yaml
 from scipy.cluster.vq import kmeans
 from tqdm import tqdm
-
 from utils.general import colorstr
 
 
@@ -15,69 +14,89 @@ def check_anchor_order(m):
     da = a[-1] - a[0]  # delta a
     ds = m.stride[-1] - m.stride[0]  # delta s
     if da.sign() != ds.sign():  # same order
-        print('Reversing anchor order')
+        print("Reversing anchor order")
         m.anchors[:] = m.anchors.flip(0)
         m.anchor_grid[:] = m.anchor_grid.flip(0)
 
 
 def check_anchors(dataset, model, thr=4.0, imgsz=640):
     # Check anchor fit to data, recompute if necessary
-    prefix = colorstr('autoanchor: ')
-    print(f'\n{prefix}Analyzing anchors... ', end='')
-    m = model.module.model[-1] if hasattr(model, 'module') else model.model[-1]  # Detect()
+    prefix = colorstr("autoanchor: ")
+    print(f"\n{prefix}Analyzing anchors... ", end="")
+    m = (
+        model.module.model[-1] if hasattr(model, "module") else model.model[-1]
+    )  # Detect()
     shapes = imgsz * dataset.shapes / dataset.shapes.max(1, keepdims=True)
     scale = np.random.uniform(0.9, 1.1, size=(shapes.shape[0], 1))  # augment scale
-    wh = torch.tensor(np.concatenate([l[:, 3:5] * s for s, l in zip(shapes * scale, dataset.labels)])).float()  # wh
+    wh = torch.tensor(
+        np.concatenate([l[:, 3:5] * s for s, l in zip(shapes * scale, dataset.labels)])
+    ).float()  # wh
 
     def metric(k):  # compute metric
         r = wh[:, None] / k[None]
-        x = torch.min(r, 1. / r).min(2)[0]  # ratio metric
+        x = torch.min(r, 1.0 / r).min(2)[0]  # ratio metric
         best = x.max(1)[0]  # best_x
-        aat = (x > 1. / thr).float().sum(1).mean()  # anchors above threshold
-        bpr = (best > 1. / thr).float().mean()  # best possible recall
+        aat = (x > 1.0 / thr).float().sum(1).mean()  # anchors above threshold
+        bpr = (best > 1.0 / thr).float().mean()  # best possible recall
         return bpr, aat
 
     bpr, aat = metric(m.anchor_grid.clone().cpu().view(-1, 2))
-    print(f'anchors/target = {aat:.2f}, Best Possible Recall (BPR) = {bpr:.4f}', end='')
+    print(f"anchors/target = {aat:.2f}, Best Possible Recall (BPR) = {bpr:.4f}", end="")
     if bpr < 0.98:  # threshold to recompute
-        print('. Attempting to improve anchors, please wait...')
+        print(". Attempting to improve anchors, please wait...")
         na = m.anchor_grid.numel() // 2  # number of anchors
-        new_anchors = kmean_anchors(dataset, n=na, img_size=imgsz, thr=thr, gen=1000, verbose=False)
+        new_anchors = kmean_anchors(
+            dataset, n=na, img_size=imgsz, thr=thr, gen=1000, verbose=False
+        )
         new_bpr = metric(new_anchors.reshape(-1, 2))[0]
         if new_bpr > bpr:  # replace anchors
-            new_anchors = torch.tensor(new_anchors, device=m.anchors.device).type_as(m.anchors)
-            m.anchor_grid[:] = new_anchors.clone().view_as(m.anchor_grid)  # for inference
-            m.anchors[:] = new_anchors.clone().view_as(m.anchors) / m.stride.to(m.anchors.device).view(-1, 1, 1)  # loss
+            new_anchors = torch.tensor(new_anchors, device=m.anchors.device).type_as(
+                m.anchors
+            )
+            m.anchor_grid[:] = new_anchors.clone().view_as(
+                m.anchor_grid
+            )  # for inference
+            m.anchors[:] = new_anchors.clone().view_as(m.anchors) / m.stride.to(
+                m.anchors.device
+            ).view(
+                -1, 1, 1
+            )  # loss
             check_anchor_order(m)
-            print(f'{prefix}New anchors saved to model. Update model *.yaml to use these anchors in the future.')
+            print(
+                f"{prefix}New anchors saved to model. Update model *.yaml to use these anchors in the future."
+            )
         else:
-            print(f'{prefix}Original anchors better than new anchors. Proceeding with original anchors.')
-    print('')  # newline
+            print(
+                f"{prefix}Original anchors better than new anchors. Proceeding with original anchors."
+            )
+    print("")  # newline
 
 
-def kmean_anchors(path='./data/coco128.yaml', n=9, img_size=640, thr=4.0, gen=1000, verbose=True):
-    """ Creates kmeans-evolved anchors from training dataset
+def kmean_anchors(
+    path="./data/coco128.yaml", n=9, img_size=640, thr=4.0, gen=1000, verbose=True
+):
+    """Creates kmeans-evolved anchors from training dataset
 
-        Arguments:
-            path: path to dataset *.yaml, or a loaded dataset
-            n: number of anchors
-            img_size: image size used for training
-            thr: anchor-label wh ratio threshold hyperparameter hyp['anchor_t'] used for training, default=4.0
-            gen: generations to evolve anchors using genetic algorithm
-            verbose: print all results
+    Arguments:
+        path: path to dataset *.yaml, or a loaded dataset
+        n: number of anchors
+        img_size: image size used for training
+        thr: anchor-label wh ratio threshold hyperparameter hyp['anchor_t'] used for training, default=4.0
+        gen: generations to evolve anchors using genetic algorithm
+        verbose: print all results
 
-        Return:
-            k: kmeans evolved anchors
+    Return:
+        k: kmeans evolved anchors
 
-        Usage:
-            from utils.autoanchor import *; _ = kmean_anchors()
+    Usage:
+        from utils.autoanchor import *; _ = kmean_anchors()
     """
-    thr = 1. / thr
-    prefix = colorstr('autoanchor: ')
+    thr = 1.0 / thr
+    prefix = colorstr("autoanchor: ")
 
     def metric(k, wh):  # compute metrics
         r = wh[:, None] / k[None]
-        x = torch.min(r, 1. / r).min(2)[0]  # ratio metric
+        x = torch.min(r, 1.0 / r).min(2)[0]  # ratio metric
         # x = wh_iou(wh, torch.tensor(k))  # iou metric
         return x, x.max(1)[0]  # x, best_x
 
@@ -88,19 +107,30 @@ def kmean_anchors(path='./data/coco128.yaml', n=9, img_size=640, thr=4.0, gen=10
     def print_results(k):
         k = k[np.argsort(k.prod(1))]  # sort small to large
         x, best = metric(k, wh0)
-        bpr, aat = (best > thr).float().mean(), (x > thr).float().mean() * n  # best possible recall, anch > thr
-        print(f'{prefix}thr={thr:.2f}: {bpr:.4f} best possible recall, {aat:.2f} anchors past thr')
-        print(f'{prefix}n={n}, img_size={img_size}, metric_all={x.mean():.3f}/{best.mean():.3f}-mean/best, '
-              f'past_thr={x[x > thr].mean():.3f}-mean: ', end='')
+        bpr, aat = (best > thr).float().mean(), (
+            x > thr
+        ).float().mean() * n  # best possible recall, anch > thr
+        print(
+            f"{prefix}thr={thr:.2f}: {bpr:.4f} best possible recall, {aat:.2f} anchors past thr"
+        )
+        print(
+            f"{prefix}n={n}, img_size={img_size}, metric_all={x.mean():.3f}/{best.mean():.3f}-mean/best, "
+            f"past_thr={x[x > thr].mean():.3f}-mean: ",
+            end="",
+        )
         for i, x in enumerate(k):
-            print('%i,%i' % (round(x[0]), round(x[1])), end=',  ' if i < len(k) - 1 else '\n')  # use in *.cfg
+            print(
+                "%i,%i" % (round(x[0]), round(x[1])),
+                end=",  " if i < len(k) - 1 else "\n",
+            )  # use in *.cfg
         return k
 
     if isinstance(path, str):  # *.yaml file
         with open(path) as f:
             data_dict = yaml.load(f, Loader=yaml.SafeLoader)  # model dict
         from utils.datasets import LoadImagesAndLabels
-        dataset = LoadImagesAndLabels(data_dict['train'], augment=True, rect=True)
+
+        dataset = LoadImagesAndLabels(data_dict["train"], augment=True, rect=True)
     else:
         dataset = path  # dataset
 
@@ -111,12 +141,14 @@ def kmean_anchors(path='./data/coco128.yaml', n=9, img_size=640, thr=4.0, gen=10
     # Filter
     i = (wh0 < 3.0).any(1).sum()
     if i:
-        print(f'{prefix}WARNING: Extremely small objects found. {i} of {len(wh0)} labels are < 3 pixels in size.')
+        print(
+            f"{prefix}WARNING: Extremely small objects found. {i} of {len(wh0)} labels are < 3 pixels in size."
+        )
     wh = wh0[(wh0 >= 2.0).any(1)]  # filter > 2 pixels
     # wh = wh * (np.random.rand(wh.shape[0], 1) * 0.9 + 0.1)  # multiply by random scale 0-1
 
     # Kmeans calculation
-    print(f'{prefix}Running kmeans for {n} anchors on {len(wh)} points...')
+    print(f"{prefix}Running kmeans for {n} anchors on {len(wh)} points...")
     s = wh.std(0)  # sigmas for whitening
     k, dist = kmeans(wh / s, n, iter=30)  # points, mean distance
     k *= s
@@ -138,17 +170,28 @@ def kmean_anchors(path='./data/coco128.yaml', n=9, img_size=640, thr=4.0, gen=10
 
     # Evolve
     npr = np.random
-    f, sh, mp, s = anchor_fitness(k), k.shape, 0.9, 0.1  # fitness, generations, mutation prob, sigma
-    pbar = tqdm(range(gen), desc=f'{prefix}Evolving anchors with Genetic Algorithm:')  # progress bar
+    f, sh, mp, s = (
+        anchor_fitness(k),
+        k.shape,
+        0.9,
+        0.1,
+    )  # fitness, generations, mutation prob, sigma
+    pbar = tqdm(
+        range(gen), desc=f"{prefix}Evolving anchors with Genetic Algorithm:"
+    )  # progress bar
     for _ in pbar:
         v = np.ones(sh)
         while (v == 1).all():  # mutate until a change occurs (prevent duplicates)
-            v = ((npr.random(sh) < mp) * npr.random() * npr.randn(*sh) * s + 1).clip(0.3, 3.0)
+            v = ((npr.random(sh) < mp) * npr.random() * npr.randn(*sh) * s + 1).clip(
+                0.3, 3.0
+            )
         kg = (k.copy() * v).clip(min=2.0)
         fg = anchor_fitness(kg)
         if fg > f:
             f, k = fg, kg.copy()
-            pbar.desc = f'{prefix}Evolving anchors with Genetic Algorithm: fitness = {f:.4f}'
+            pbar.desc = (
+                f"{prefix}Evolving anchors with Genetic Algorithm: fitness = {f:.4f}"
+            )
             if verbose:
                 print_results(k)
